create a new script for risk level
